feat: expose useCustomControl
