+++ packages/core/src/useCustomField/index.ts
@@ -1,1 +1,2 @@
 export * from './useCustomField';
+export * from './useCustomControl';
